refactor: MenuGroupRequest 생성
