+++ src/main/java/kitchenpos/application/MenuGroupService.java
@@ -1,11 +1,11 @@
 package kitchenpos.application;
 
+import java.util.List;
 import kitchenpos.dao.MenuGroupDao;
 import kitchenpos.domain.MenuGroup;
+import kitchenpos.ui.dto.MenuGroupRequest;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-
-import java.util.List;
 
 @Service
 public class MenuGroupService {
@@ -16,8 +16,8 @@ public class MenuGroupService {
     }
 
     @Transactional
-    public MenuGroup create(final MenuGroup menuGroup) {
-        return menuGroupDao.save(menuGroup);
+    public MenuGroup create(final MenuGroupRequest request) {
+        return menuGroupDao.save(request.toMenuGroup());
     }
 
     public List<MenuGroup> list() {
